backend/requirements.txt
fastapi
uvicorn[standard]
prisma
python-dotenv
pydantic
Jinja2
psycopg2-binary # For PostgreSQL
# Add any other specific backend dependencies if needed 